requirements.txt
torch
torchvision
Pillow
line_profiler
